Add MyST yml header to each notebook (#27)
Add a yml header to the first cell of notebooks including title and
authors. Improve list of authors in `myst.yml`. Add Lindsey to the list
of authors. Add UBC as an affiliation with id so we can reuse it on each
author. Add ORCIDs. Add jupyterlab-myst to environment.yml to render
those cells in Jupyter Lab.

Fixes #26